+++ src/util/helper.rs
@@ -195,12 +195,6 @@ pub fn fit_view_box(size: Size, vb: &usvg::ViewBox) -> usvg::Size {
     }
 }
 
-/// Calculate the scale ratio of a DPI value.
-/// Turns a DPI in a dot per user unit (1/72nd of an inch by default)
-pub fn dpi_ratio(dpi: f32) -> f32 {
-    dpi / 72.0
-}
-
 /// Compress data using the deflate algorithm.
 pub fn deflate(data: &[u8]) -> Vec<u8> {
     const COMPRESSION_LEVEL: u8 = 6;