+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added support for text embedding.
+- The `convert_str` method has been removed. You should now always convert your SVG string into a `usvg` 
+ tree yourself.
+- The `convert_tree` method has been renamed into `to_pdf`, and now requires you to provide the fontdb
+ used for the `usvg` tree.
+- `convert_tree_into` has been renamed into `to_chunk` and now returns an independent chunk as well
+ as the object ID of the SVG.
+
+- TODO: The CLI options have been (temporarily) removed. They will be readded before the next release.
+- TODO: Add tests for CLI and svg options
+- TODO: Add CLI option to convert text to paths.
 
 ### Changed
 - Bumped resvg to v0.40.